// src/app/generated/model/purchaseOrderLineRequest.ts
/**
 * BC Construction Services API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * One line item on a purchase order
 */
export interface PurchaseOrderLineRequest { 
    /**
     * Identifier of the item being ordered
     */
    itemId: number;
    /**
     * Quantity being ordered
     */
    quantity: number;
    /**
     * Optional free-text notes about this line
     */
    notes?: string;
}